refactor: move copyDockerRun to utils/ui, and fix crypto and auth imports
