+++ src/lib/utils/ui/clipboard.ts
@@ -1,3 +1,5 @@
+import { toast } from 'svelte-sonner';
+
 export async function copyTextToClipboard(text: string): Promise<boolean> {
 	try {
 		await navigator.clipboard.writeText(text);
@@ -6,4 +8,25 @@ export async function copyTextToClipboard(text: string): Promise<boolean> {
 		console.error('Failed to copy text: ', error);
 		return false;
 	}
+}
+
+export async function copyDockerRunCommand(imageFullName: string, tagName: string, registryURL: string): Promise<void> {
+	let registryHost = '';
+	try {
+		const url = new URL(registryURL);
+		registryHost = url.host;
+	} catch (e) {
+		// Fallback if URL parsing fails
+		registryHost = registryURL.replace(/^https?:\/\//, '');
+	}
+
+	const dockerRunCmd = `docker run ${registryHost}/${imageFullName}:${tagName}`;
+
+	copyTextToClipboard(dockerRunCmd).then((success) => {
+		if (success) {
+			toast.success('Docker Run command copied to clipboard');
+		} else {
+			toast.error('Failed to copy Docker Run command');
+		}
+	});
 }
