+++ src/routes/details/[repo]/[image]/[tag]/+page.svelte
@@ -15,6 +15,7 @@
 	import { env } from '$env/dynamic/public';
 	import * as AlertDialog from '$lib/components/ui/alert-dialog/index.js';
 	import { onMount, onDestroy } from 'svelte';
+	import { copyDockerRunCommand } from '$lib/utils/ui';}
 
 	export let data: PageData;
 
@@ -52,26 +53,26 @@
 		}
 	});
 
-	async function copyDockerRunCommand() {
-		let registryHost = '';
-		try {
-			const url = new URL(env.PUBLIC_REGISTRY_URL);
-			registryHost = url.host;
-		} catch (e) {
-			// Fallback if URL parsing fails
-			registryHost = env.PUBLIC_REGISTRY_URL.replace(/^https?:\/\//, '');
-		}
+	// async function copyDockerRunCommand() {
+	// 	let registryHost = '';
+	// 	try {
+	// 		const url = new URL(env.PUBLIC_REGISTRY_URL);
+	// 		registryHost = url.host;
+	// 	} catch (e) {
+	// 		// Fallback if URL parsing fails
+	// 		registryHost = env.PUBLIC_REGISTRY_URL.replace(/^https?:\/\//, '');
+	// 	}
 
-		const dockerRunCmd = `docker run ${registryHost}/${data.imageFullName}:${currentTag.name}`;
+	// 	const dockerRunCmd = `docker run ${registryHost}/${data.imageFullName}:${currentTag.name}`;
 
-		copyTextToClipboard(dockerRunCmd).then((success) => {
-			if (success) {
-				toast.success('Docker Run command copied to clipboard');
-			} else {
-				toast.error('Failed to copy Docker Run command');
-			}
-		});
-	}
+	// 	copyTextToClipboard(dockerRunCmd).then((success) => {
+	// 		if (success) {
+	// 			toast.success('Docker Run command copied to clipboard');
+	// 		} else {
+	// 			toast.error('Failed to copy Docker Run command');
+	// 		}
+	// 	});
+	// }
 
 	async function deleteTagBackend(name: string, digest: string) {
 		if (!digest) {
@@ -221,7 +222,7 @@
 								<div class="flex flex-col gap-2">
 									<p class="text-sm text-muted-foreground font-mono">{currentTag.metadata?.indexDigest || ''}</p>
 									<!-- Add the Docker Run Command button here -->
-									<Button variant="outline" size="sm" class="gap-2 mt-5 w-fit" onclick={copyDockerRunCommand}>
+									<Button variant="outline" size="sm" class="gap-2 mt-5 w-fit" onclick={copyDockerRunCommand(data.imageFullName, currentTag.name, env.PUBLIC_REGISTRY_URL)}>
 										<Terminal class="h-4 w-4" />
 										Copy Docker Run Command
 									</Button>
